requirements.txt
requests
pymongo>=4.6.0
pypdf>=3.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4
sqlalchemy<2.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pymupdf>=1.23.0
sentence-transformers>=2.2.2
rasa-sdk==3.6.2
